gcc 14 fixes

+++ sql/table_expression_operations.cc
@@ -544,12 +544,12 @@ bind(SqlBindingScope & context, const ProgressFunc & onProgress) const
     
     size_t steps = args.size();
     ProgressState joinState(100);
-    auto stepProgress = [&](uint step, const ProgressState & state) {
+    auto stepProgress = [&](size_t step, const ProgressState & state) {
         joinState = (100 / steps * state.count / *state.total) + (100 / steps * step);
         return onProgress(joinState);
     };
 
-    for (uint i = 0; i < steps; ++i) {
+    for (size_t i = 0; i < steps; ++i) {
         auto & arg = args[i];
         auto & combinedProgress = onProgress ? std::bind(stepProgress, i, _1) : onProgress;
         boundArgs.push_back(arg->bind(context, combinedProgress));
